Add files via upload
In this data analysis project, I used SQL and Python to analyze and understand the dataset. The analysis involved solving a total of 13 questions, ranging from basic to advanced levels. I applied Matplotlib and Seaborn libraries to create visualizations wherever necessary, which helped in identifying trends and insights more effectively.
Through this project, I gained a deeper understanding of how to clean, query, and visualize data, and how to draw meaningful conclusions from it using both SQL and Python tools.